Add support for currency in FirstData purchases & refunds.
Use the 'currency' value specified in the options hash param, if available, otherwise use default_currency ('USD').  Although not recommended, if the amount passed into purchase or refund requests is a Money object, and a currency isn't included in the options hash param, the currency is extracted from the amount Money object.

Unit & remote tests have been added to verify that specifying a currency in the options hash param will result in the response confirming the transaction.

NOTE: with my test credentials, the remote test that checks the AVS result code returns '4' instead of '1'.  As this test is unrelated to these changes, I'm assuming the cause is related to the test credentials (or the test was previously broken).